Fixed 32 bit build [ci patch inc] (#17)

+++ src/Emulator/Core/MemoryController/MemoryControllerST.hpp
@@ -18,7 +18,7 @@ namespace HyperCPU {
       if (!memory)
         throw std::runtime_error("Failed to allocate memory!");
     }
-    inline std::uint8_t Fetch8(std::size_t& ptr) override {
+    inline std::uint8_t Fetch8(std::uint64_t& ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint8_t) - 1 < total_mem);
       std::uint8_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint8_t));
@@ -26,7 +26,7 @@ namespace HyperCPU {
       return data;
     }
 
-    inline std::uint16_t Fetch16(std::size_t& ptr) override {
+    inline std::uint16_t Fetch16(std::uint64_t& ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint16_t) - 1 < total_mem);
       std::uint16_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint16_t));
@@ -34,7 +34,7 @@ namespace HyperCPU {
       return data;
     }
 
-    inline std::uint32_t Fetch32(std::size_t& ptr) override {
+    inline std::uint32_t Fetch32(std::uint64_t& ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint32_t) - 1 < total_mem);
       std::uint32_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint32_t));
@@ -42,7 +42,7 @@ namespace HyperCPU {
       return data;
     }
 
-    inline std::uint64_t Fetch64(std::size_t& ptr) override {
+    inline std::uint64_t Fetch64(std::uint64_t& ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint64_t) - 1 < total_mem);
       std::uint64_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint64_t));
@@ -50,50 +50,50 @@ namespace HyperCPU {
       return data;
     }
 
-    inline std::uint8_t Read8(std::size_t ptr) override {
+    inline std::uint8_t Read8(std::uint64_t ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint8_t) - 1 < total_mem);
       std::uint8_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint8_t));
       return data;
     }
 
-    inline std::uint16_t Read16(std::size_t ptr) override {
+    inline std::uint16_t Read16(std::uint64_t ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint16_t) - 1 < total_mem);
       std::uint16_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint16_t));
       return data;
     }
 
-    inline std::uint32_t Read32(std::size_t ptr) override {
+    inline std::uint32_t Read32(std::uint64_t ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint32_t) - 1 < total_mem);
       std::uint32_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint32_t));
       return data;
     }
 
-    inline std::uint64_t Read64(std::size_t ptr) override {
+    inline std::uint64_t Read64(std::uint64_t ptr) override {
       mem_ctlr_assert(ptr + sizeof(std::uint64_t) - 1 < total_mem);
       std::uint64_t data;
       memcpy(&data, &memory[ptr], sizeof(std::uint64_t));
       return data;
     }
 
-    inline void Load8(std::size_t ptr, std::uint8_t data) override {
+    inline void Load8(std::uint64_t ptr, std::uint8_t data) override {
       mem_ctlr_assert(ptr + sizeof(std::uint8_t) - 1 < total_mem);
       memcpy(&memory[ptr], &data, sizeof(std::uint8_t));
     }
 
-    inline void Load16(std::size_t ptr, std::uint16_t data) override {
+    inline void Load16(std::uint64_t ptr, std::uint16_t data) override {
       mem_ctlr_assert(ptr + sizeof(std::uint16_t) - 1 < total_mem);
       memcpy(&memory[ptr], &data, sizeof(std::uint16_t));
     }
 
-    inline void Load32(std::size_t ptr, std::uint32_t data) override {
+    inline void Load32(std::uint64_t ptr, std::uint32_t data) override {
       mem_ctlr_assert(ptr + sizeof(std::uint32_t) - 1 < total_mem);
       memcpy(&memory[ptr], &data, sizeof(std::uint32_t));
     }
 
-    inline void Load64(std::size_t ptr, std::uint64_t data) override {
+    inline void Load64(std::uint64_t ptr, std::uint64_t data) override {
       mem_ctlr_assert(ptr + sizeof(std::uint64_t) - 1 < total_mem);
       memcpy(&memory[ptr], &data, sizeof(std::uint64_t));
     }
